+++ src/_build/header.py
@@ -352,9 +352,14 @@ def generate_header(h_dt,cm):
 			if (st is False):
 				o+=b","
 			st=False
-			o+=a[:-len(a.split(b" ")[-1])].strip()
-			if (len(a[:-len(a.split(b" ")[-1])].strip())>0):
+			if (a==b"..."):
+				o+=b"..."
 				void=False
+				break
+			else:
+				o+=a[:-len(a.split(b" ")[-1])].strip()
+				if (len(a[:-len(a.split(b" ")[-1])].strip())>0):
+					void=False
 		if (void):
 			o+=b"void"
 		o+=b");"